requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
openpyxl==3.1.2
Werkzeug==3.0.1
